requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
numpy==1.26.4
open3d==0.18.0
Pillow==10.3.0
Requests==2.31.0
torch==2.2.2
transformers==4.40.0
Werkzeug==2.2.3